Djsopnet views: update for segmentation schemas
Update views and CATSOP results widget to restore basic functionality
with segmentation stack-specific schemas. Currently the CATSOP widget
defaults to choosing the first SegmentationConfiguration available for
the (stack, project) tuple and uses the membrane SegmentStack from that
configuration.

Since segmentation tables no longer have Django ORM models, take the
cursor-to-namedtuple approach previously used to query slices and apply
it also to segments, blocks and cores.

Most non-deprecated views now have routes including project_id and
segmentation_stack_id, rather than stack_id. Since a SegmentationStack
implies a Project, project_id is present only for compatibility with
access control decorators.

Note that none of the views set a schema search path for the connection
cursor because of fault- and concurrency-tolerance, so instead prepend
the appropriate segmentation schema name to all table references.

Correct the type of the mistake_id column for corrections, which
references segments, to be bigint, via a migration. This migration
represents the correct approach for making changes to the unmanaged
segmentation schemas. Correction mistake_id used to refer to specific
segment_solutions, but since those no longer have a separate primary key
and the solution can be inferred, mistake_id now refers directly to the
segment that was incorrectly in the solution.

+++ django/applications/djsopnet/sql/segmentation_schema.sql
@@ -263,7 +263,7 @@ CREATE TABLE constraint_segment_relation (
 
 CREATE TABLE correction (
   constraint_id integer NOT NULL REFERENCES solution_constraint(id),
-  mistake_id integer NOT NULL REFERENCES segment(id),
+  mistake_id bigint NOT NULL REFERENCES segment(id),
   PRIMARY KEY (constraint_id, mistake_id),
   CHECK (false) NO INHERIT -- prevent any rows populating this table
 ) WITH (
